[#31] Schedule Notifications are now a subclass of the base Notification class
