+++ app/src/main/java/com/example/bigowlapp/model/Notification.java
@@ -9,13 +9,8 @@ public class Notification {
 
     @DocumentId
     private String uId;
-    private String senderUId;
-    private String receiverUId;
-    private String groupUId;
     private String type;
-    private Timestamp timeRead;
-    private Timestamp timeSend;
-    private UserScheduleResponse senderResponse;
+    private Timestamp time;
 
     public Notification() {
     }
@@ -29,61 +24,10 @@ public class Notification {
         this.type = type;
     }
 
-    public Notification(String uId, String type, Timestamp timeRead) {
+    public Notification(String uId, String type, Timestamp time) {
         this.uId = uId;
         this.type = type;
-        this.timeRead = timeRead;
-    }
-
-    public Notification(String uId, String senderUId, String receiverUId, String groupUId, String type, Timestamp timeRead, Timestamp timeSend, UserScheduleResponse senderResponse) {
-        this.uId = uId;
-        this.senderUId = senderUId;
-        this.receiverUId = receiverUId;
-        this.groupUId = groupUId;
-        this.type = type;
-        this.timeRead = timeRead;
-        this.timeSend = timeSend;
-        this.senderResponse = senderResponse;
-    }
-
-    public UserScheduleResponse getSenderResponse() {
-        return senderResponse;
-    }
-
-    public void setSenderResponse(UserScheduleResponse senderResponse) {
-        this.senderResponse = senderResponse;
-    }
-
-    public String getSenderUId() {
-        return senderUId;
-    }
-
-    public void setSenderUId(String senderUId) {
-        this.senderUId = senderUId;
-    }
-
-    public String getReceiverUId() {
-        return receiverUId;
-    }
-
-    public void setReceiverUId(String receiverUId) {
-        this.receiverUId = receiverUId;
-    }
-
-    public String getGroupUId() {
-        return groupUId;
-    }
-
-    public void setGroupUId(String groupUId) {
-        this.groupUId = groupUId;
-    }
-
-    public Timestamp getTimeSend() {
-        return timeSend;
-    }
-
-    public void setTimeSend(Timestamp timeSend) {
-        this.timeSend = timeSend;
+        this.time = time;
     }
 
     public String getuId() {
@@ -102,11 +46,11 @@ public class Notification {
         this.type = type;
     }
 
-    public Timestamp getTimeRead() {
-        return timeRead;
+    public Timestamp getTime() {
+        return time;
     }
 
-    public void setTimeRead(Timestamp timeRead) {
-        this.timeRead = timeRead;
+    public void setTime(Timestamp time) {
+        this.time = time;
     }
 }
