+++ app/src/main/java/com/example/bigowlapp/repository/NotificationRepository.java
@@ -1,12 +1,12 @@
 package com.example.bigowlapp.repository;
 
-import android.opengl.Visibility;
 import android.util.Log;
 
 import androidx.annotation.VisibleForTesting;
 import androidx.lifecycle.MutableLiveData;
 
 import com.example.bigowlapp.model.Notification;
+import com.example.bigowlapp.model.ScheduleRequest;
 import com.example.bigowlapp.model.SupervisionRequest;
 import com.example.bigowlapp.utils.Constants;
 import com.google.firebase.firestore.CollectionReference;
@@ -31,7 +31,7 @@ public class NotificationRepository extends Repository<Notification> {
     }
 
     public MutableLiveData<List<SupervisionRequest>> getListOfSupervisionRequestByAttribute(String attribute, String attrValue,
-                                                                                  Class<? extends SupervisionRequest> tClass) {
+                                                                                            Class<? extends SupervisionRequest> tClass) {
         MutableLiveData<List<SupervisionRequest>> listOfTData = new MutableLiveData<>();
         collectionReference.whereEqualTo(attribute, attrValue)
                 .whereEqualTo("type", Constants.SUPERVISION_TYPE)
@@ -69,14 +69,15 @@ public class NotificationRepository extends Repository<Notification> {
     }
 
     Class<? extends Notification> getClassFromType(String type) {
-        if (type == null) {
-            return Notification.class;
-        }
+        switch (String.valueOf(type)) {
+            case Constants.SUPERVISION_TYPE:
+                return SupervisionRequest.class;
 
-        if (type.equals(Constants.SUPERVISION_TYPE)) {
-            return SupervisionRequest.class;
-        } else {
-            return Notification.class;
+            case Constants.SCHEDULE_TYPE:
+                return ScheduleRequest.class;
+
+            default:
+                return Notification.class;
         }
     }
 }
