+++ app/src/main/java/com/example/bigowlapp/utils/Constants.java
@@ -3,6 +3,7 @@ package com.example.bigowlapp.utils;
 public interface Constants {
 
     String SUPERVISION_TYPE = "supervisionRequest";
+    String SCHEDULE_TYPE = "scheduleRequest";
 
     // Splash screen time out in milliseconds
     int SPLASH_DURATION = 2000;
